feat(frontend): add chart in history page
